Consolidate into single script
